Move semester display to utils
